requirements.txt
asgiref==3.11.0
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==6.0
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
Flask-SocketIO==5.6.0
greenlet==3.3.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pdf2image==1.17.0
pillow==12.1.0
pytesseract==0.3.13
python-engineio==4.13.0
python-socketio==5.16.0
simple-websocket==1.1.0
sqlparse==0.5.4
tzdata==2025.2
Werkzeug==3.1.5
wsproto==1.3.2
